refactor: BookItemsとDndListのフィルターUIとメッセージ定義の重複を共通化
